Addition of the 160909 Compilation
Lots of Stuff, check it out - should be all OK.

Had to return to the Britain.txt & Additional_UoB.txt instead of the new
Domestic and Foreign files as hadn't accounted for Matej's
reorganisation. Files seem the same, and the id's are the same format,
but all the new content is on the restored files, so that's what I'm
proposing for now.